Extract client compilation to separate class
